enfileTaTuque-backend/server/data_fetching/requirements.txt
azure-core==1.26.0
azure-storage-blob==12.14.1
pandas==1.4.4
python-dotenv==0.21.0
requests==2.28.1
